requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.4.5.1
colorama==0.4.3
dnspython==1.16.0
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
mysql-connector-python==8.0.19
packaging==20.3
pluggy==0.13.1
protobuf==3.6.1
py==1.8.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
six==1.14.0
typed-ast==1.4.1
wcwidth==0.1.9
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
